[追加]MoveクラスにAction型を戻り値とするメソッドを追加。全方向分
