+++ Main/src/character/type/compatibility/WaterTypeCompatibilityCheck.java
@@ -1,6 +1,6 @@
-package type.compatibility;
+package character.type.compatibility;
 
-import type.Types;
+import character.type.Types;
 
 class WaterTypeCompatibilityCheck implements CompatibilityCheck {
     @Override
