+++ Main/src/field/point/move/DownAction.java
@@ -1,11 +1,11 @@
-package field.move;
+package field.point.move;
 
 import field.point.Point;
 
-class RightAction implements Action{
+class DownAction implements Action{
     @Override
     public Point go(Point point) {
-        return Point.init(point.getRow(), point.getColumn() + MoveLength.normal);
+        return Point.init(point.getRow() - MoveLength.normal, point.getColumn());
     }
 
     @Override
