+++ Main/src/field/point/move/MoveLength.java
@@ -1,4 +1,4 @@
-package field.move;
+package field.point.move;
 
 class MoveLength {
     static final int normal = 1;
